pyproject.toml
[project]
name = "mistral-learn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.37.0",
    "fsrs>=6.3.0",
    "gradio>=6.8.0",
    "mistralai>=1.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "wikipedia-api>=0.9.0",
    "youtube-transcript-api>=1.2.4",
]
